lib: call the two s16 flag tests through their prototypes
ActorsShared8016974c and ActorsShared8016945c were declared s32 and called
without their Task argument, on the belief that a prototype would add a
register copy; it does not, since the Task is still in $a0. Gp_LcgState is
one u32, so the member that declared it s32 and cast it back no longer does.

+++ src/lib/actors_shared_80169dbc.c
@@ -2,7 +2,7 @@
 
 #include "actors/actors_shared_80169dbc.h"
 
-extern s32 Gp_LcgState;
+extern u32 Gp_LcgState;
 
 void ActorsShared80169dbc(Task* arg0)
 {
@@ -16,7 +16,7 @@ void ActorsShared80169dbc(Task* arg0)
     /* Rolling the LCG through the global rather than an m2c temporary is what
      * hoists its `lw` above the field stores; see DECOMPILATION_LEARNINGS.md. */
     Gp_LcgState     = Gp_LcgState * 5 + 0x71357911;
-    work->field_446 = (((u32)Gp_LcgState >> 16) & 0x3F) + 0x60;
+    work->field_446 = ((Gp_LcgState >> 16) & 0x3F) + 0x60;
     work->field_412 = 0;
     work->field_422 = work->field_422 + 1;
 }
